LayoutTests/fast/regions/css-regions-disabled-expected.txt

Test that parsing of css regions related properties is disabled by default.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testWebKitFlowInto("-webkit-flow-into: auto") is ""
PASS testWebKitFlowInto("-webkit-flow-into: first-flow") is ""
PASS testComputedStyleWebKitFlowInto("auto") is "auto"
PASS testComputedStyleWebKitFlowInto("first-flow") is "auto"
PASS testWebKitFlowFrom("-webkit-flow-from: first-flow") is ""
PASS testWebKitFlowFrom("-webkit-flow-from: none") is ""
PASS testComputedStyleWebKitFlowFrom("first-flow") is "none"
PASS testComputedStyleWebKitFlowFrom("none") is "none"
PASS testWebKitRegionOverflow("-webkit-region-overflow: auto") is ""
PASS testWebKitRegionOverflow("-webkit-region-overflow: break") is ""
PASS testComputedStyleWebKitRegionOverflow("auto") is "auto"
PASS testComputedStyleWebKitRegionOverflow("break") is "auto"
PASS webkitRegionRuleIndex is -1
PASS successfullyParsed is true

TEST COMPLETE
